refactor: rename to App.jsx
